Restructured code in line with recommendations
https://github.com/golang-standards/project-layout/tree/master\?tab\=readme-ov-file\#cmd
