+++ tile-id-api/internal/handler/quadkey/quadkey_test.go
@@ -7,7 +7,7 @@ import (
 	"reflect"
 	"testing"
 
-	"github.com/captaincoordinates/tile-id-api/tile-id-api/constants"
+	"github.com/captaincoordinates/tile-id-api/tile-id-api/internal/constants"
 	"github.com/sirupsen/logrus"
 )
