+++ tile-id-api/internal/log/log.go
@@ -3,7 +3,7 @@ package log
 import (
 	"fmt"
 
-	"github.com/captaincoordinates/tile-id-api/tile-id-api/constants"
+	"github.com/captaincoordinates/tile-id-api/tile-id-api/internal/constants"
 	"github.com/sirupsen/logrus"
 )
