+++ tile-id-api/internal/params/params.go
@@ -8,8 +8,8 @@ import (
 	"regexp"
 	"strconv"
 
-	"github.com/captaincoordinates/tile-id-api/tile-id-api/constants"
-	"github.com/captaincoordinates/tile-id-api/tile-id-api/handler"
+	"github.com/captaincoordinates/tile-id-api/tile-id-api/internal/constants"
+	"github.com/captaincoordinates/tile-id-api/tile-id-api/internal/handler"
 	"github.com/gorilla/mux"
 	"github.com/sirupsen/logrus"
 )
