+++ tile-id-api/internal/params/params_test.go
@@ -7,8 +7,8 @@ import (
 	"net/url"
 	"testing"
 
-	"github.com/captaincoordinates/tile-id-api/tile-id-api/constants"
-	"github.com/captaincoordinates/tile-id-api/tile-id-api/handler"
+	"github.com/captaincoordinates/tile-id-api/tile-id-api/internal/constants"
+	"github.com/captaincoordinates/tile-id-api/tile-id-api/internal/handler"
 )
 
 func TestIntPathParamsValid(t *testing.T) {
